+++ src/LogExpert/Extensions/MenuItemExtension.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace LogExpert.Extensions
+{
+    public class ExtendedMenuItemRenderer : ToolStripProfessionalRenderer
+    {
+        public ExtendedMenuItemRenderer() : base(new MenuSelectedColors()) { }
+    }
+
+    public class MenuSelectedColors : ProfessionalColorTable
+    {
+        public override Color MenuItemPressedGradientBegin
+        {
+            get { return LogExpert.Config.ColorMode.MenuBackgroundColor; }
+        }
+
+        public override Color MenuItemPressedGradientEnd
+        {
+            get { return LogExpert.Config.ColorMode.MenuBackgroundColor; }
+        }
+    }
+}
